Module aws_sdk_mediastore::error

Modules

container_in_use_exception

See ContainerInUseException

container_not_found_exception

See ContainerNotFoundException

cors_policy_not_found_exception

See CorsPolicyNotFoundException

internal_server_error

See InternalServerError

limit_exceeded_exception

See LimitExceededException

policy_not_found_exception

See PolicyNotFoundException

Structs

ContainerInUseException

The container that you specified in the request already exists or is being

ContainerNotFoundException

The container that you specified in the request does not exist.

CorsPolicyNotFoundException

The CORS policy that you specified in the request does not exist.

CreateContainerError

DeleteContainerError

DeleteContainerPolicyError

DeleteCorsPolicyError

DeleteLifecyclePolicyError

DeleteMetricPolicyError

DescribeContainerError

GetContainerPolicyError

GetCorsPolicyError

GetLifecyclePolicyError

GetMetricPolicyError

InternalServerError

The service is temporarily unavailable.

LimitExceededException

A service limit has been exceeded.

ListContainersError

ListTagsForResourceError

PolicyNotFoundException

The policy that you specified in the request does not exist.

PutContainerPolicyError

PutCorsPolicyError

PutLifecyclePolicyError

PutMetricPolicyError

StartAccessLoggingError

StopAccessLoggingError

TagResourceError

UntagResourceError

Enums

CreateContainerErrorKind

DeleteContainerErrorKind

DeleteContainerPolicyErrorKind

DeleteCorsPolicyErrorKind

DeleteLifecyclePolicyErrorKind

DeleteMetricPolicyErrorKind

DescribeContainerErrorKind

GetContainerPolicyErrorKind

GetCorsPolicyErrorKind

GetLifecyclePolicyErrorKind

GetMetricPolicyErrorKind

ListContainersErrorKind

ListTagsForResourceErrorKind

PutContainerPolicyErrorKind

PutCorsPolicyErrorKind

PutLifecyclePolicyErrorKind

PutMetricPolicyErrorKind

StartAccessLoggingErrorKind

StopAccessLoggingErrorKind

TagResourceErrorKind

UntagResourceErrorKind