Module aws_sdk_mediastore::model

Modules

container

See Container

cors_rule

See CorsRule

metric_policy

See MetricPolicy

metric_policy_rule

See MetricPolicyRule

tag

See Tag

Structs

Container

This section describes operations that you can perform on an AWS Elemental MediaStore

CorsRule

A rule for a CORS policy. You can add up to 100 rules to a CORS policy. If more than

MetricPolicy

The metric policy that is associated with the container. A metric policy allows AWS Elemental MediaStore to send metrics to Amazon CloudWatch. In the policy, you must indicate whether you want MediaStore to send container-level metrics. You can also include rules to define groups of objects that you want MediaStore to send object-level metrics for.

MetricPolicyRule

A setting that enables metrics at the object level. Each rule contains an object group and an object group name. If the policy includes the MetricPolicyRules parameter, you must include at least one rule. Each metric policy can include up to five rules by default. You can also request a quota increase to allow up to 300 rules per policy.

Tag

A collection of tags associated with a container. Each tag consists of a key:value pair, which can be anything you define. Typically, the tag key

Enums

ContainerLevelMetrics

ContainerStatus

MethodName