Module aws_sdk_mediastore::operation

Structs

CreateContainer

Creates a storage container to hold objects. A container is similar to a bucket in

DeleteContainer

Deletes the specified container. Before you make a DeleteContainer

DeleteContainerPolicy

Deletes the access policy that is associated with the specified container.

DeleteCorsPolicy

Deletes the cross-origin resource sharing (CORS) configuration information that is

DeleteLifecyclePolicy

Removes an object lifecycle policy from a container. It takes up to 20 minutes for the change to take effect.

DeleteMetricPolicy

Deletes the metric policy that is associated with the specified container. If there is no metric policy associated with the container, MediaStore doesn't send metrics to CloudWatch.

DescribeContainer

Retrieves the properties of the requested container. This request is commonly used to

GetContainerPolicy

Retrieves the access policy for the specified container. For information about the

GetCorsPolicy

Returns the cross-origin resource sharing (CORS) configuration information that is

GetLifecyclePolicy

Retrieves the object lifecycle policy that is assigned to a container.

GetMetricPolicy

Returns the metric policy for the specified container.

ListContainers

Lists the properties of all containers in AWS Elemental MediaStore.

ListTagsForResource

Returns a list of the tags assigned to the specified container.

PutContainerPolicy

Creates an access policy for the specified container to restrict the users and

PutCorsPolicy

Sets the cross-origin resource sharing (CORS) configuration on a container so that

PutLifecyclePolicy

Writes an object lifecycle policy to a container. If the container already has an object lifecycle policy, the service replaces the existing policy with the new policy. It takes up to 20 minutes for the change to take effect.

PutMetricPolicy

The metric policy that you want to add to the container. A metric policy allows AWS Elemental MediaStore to send metrics to Amazon CloudWatch. It takes up to 20 minutes for the new policy to take effect.

StartAccessLogging

Starts access logging on the specified container. When you enable access logging on a container, MediaStore delivers access logs for objects stored in that container to Amazon CloudWatch Logs.

StopAccessLogging

Stops access logging on the specified container. When you stop access logging on a container, MediaStore stops sending access logs to Amazon CloudWatch Logs. These access logs are not saved and are not retrievable.

TagResource

Adds tags to the specified AWS Elemental MediaStore container. Tags are key:value pairs that you can associate with AWS resources. For example, the

UntagResource

Removes tags from the specified container. You can specify one or more tags to remove.